columnarhead: target slab, wired into a real end-to-end Head
Builds the target slab design doc §3.1 describes (targetFields uint32
symbol refs per target: cluster/namespace/pod/container/node/job) and
wires it together with SeriesStore and a new live symbol interner
into Head - GetOrCreateSeries(TargetLabels, metricName, localLabel),
the first thing in this package that looks like something
storage.Appender.Append could actually call. Not wired into Cortex's
tsdbStore interface (Phase 1's work) yet - that's separate, later
work.

Found and fixed a real correctness issue before it could bite in
production: SeriesStore.targetID was uint16, sized against every
benchmark's static snapshot (25,000 targets, comfortably under
65,536). A live head accumulates targets indefinitely as pods churn -
unlike metric names or "le"-style local labels, which come from a
bounded vocabulary, every new pod is a new target with no upper bound
over uptime. 65,536 target creations is realistic within weeks on a
high-churn fleet, not a theoretical edge case. Widened targetID to
uint32; nameID/localRef stay uint16 since that risk is real but much
smaller (bounded vocabulary, doesn't grow with churn). Head also
fails loudly (ErrTooManySymbols) rather than silently truncating if a
metric name or local label ever exceeds uint16 range.

The live path uses plain Go maps (targetIndex, seriesIndex,
liveInterner's index), not the MPHF/SymbolTable built two commits
ago: those are static with no Insert, and real ingestion sees new pod
names continuously. liveInterner (interner.go) is the honest live
counterpart - explicitly documented as paying the ~31-38 B/entry cost
bench/00/bench/03 already measured for a Go map, not the compact MPHF
representation. This is design doc §3.3's "small dynamic overlay for
series created since the last build"; the rebuild-at-compaction path
that would periodically fold this into a fresh static SymbolTable and
reset it does not exist yet.

Real, honest, end-to-end measurement (TestHeadAtScale, 500k series /
2,500 targets / 2,917 symbols, same k8s-shaped workload used
throughout): 184.9 B/series total, including every live dedup map's
real cost - not measured anywhere before this, because nothing
before this actually wired the maps in. Correctly achieves the design
doc's own claimed 200:1 target-sharing ratio.

This materially corrects every ratio quoted elsewhere in
CHECKLIST.md. Prior numbers (SeriesStore-alone ~152-176 B/series, the
post-compaction projections of ~9x/~17-19x vs the real head) were all
measured or projected without live dedup indexes a running system
actually needs - they describe the compacted, post-MPHF-rebuild
state, which doesn't exist as working code yet. The honest number for
"what does this cost right now, running" is 184.9 B/series, ~3.5x vs
the real head's 648 B/series - a real, positive, working result, but
meaningfully more modest than the compacted projections.

Verified NOT caused by this work: go test -race ./pkg/ingester/...
(the parent package) shows a pre-existing data race in
pkg/cortexpb.CopyLabels / Ingester.Push, reproduced identically on
the clean, already-committed tree via git stash + re-run before any
of this session's changes. Not introduced by, or related to, this
commit - noted only so it isn't later mistaken for a regression here.

go test -race, go vet, goimports, and golangci-lint all clean on
columnarhead.

+++ pkg/ingester/columnarhead/memory_test.go
@@ -26,7 +26,7 @@ func TestMemoryPerSeries(t *testing.T) {
 	s := NewSeriesStore(n)
 	refs := make([]uint32, n)
 	for i := 0; i < n; i++ {
-		refs[i] = s.Create(uint16(i/seriesPerTgt), uint16(i%400), uint16(i%12))
+		refs[i] = s.Create(uint32(i/seriesPerTgt), uint16(i%400), uint16(i%12))
 	}
 
 	ts := int64(1700000000000)
@@ -112,7 +112,7 @@ func TestMemoryPerSeries_Staggered(t *testing.T) {
 	for wave := 0; wave < waves; wave++ {
 		for w := 0; w < perWave; w++ {
 			i := wave*perWave + w
-			ref := s.Create(uint16(i/seriesPerTgt), uint16(i%400), uint16(i%12))
+			ref := s.Create(uint32(i/seriesPerTgt), uint16(i%400), uint16(i%12))
 			refs = append(refs, ref)
 		}
 		// Every series alive so far gets appended to this wave, same as real ingestion
